`size_t` in `mojo::DataPipe[Consumer|Producer]Handle`: /components.
This CL makes use of the new `size_t`-based APIs in the remaining
directories like `//components`, `//extensions`, or `//storage`.
See https://crrev.com/c/5414148 for motivation.

This CL may have introduced new integer conversions - in those cases the
CL attempts to have well-defined behavior (e.g. using
`base::checked_cast`).  OTOH, this CL did *not* attempt to fix old
integer operations (e.g. old `int` => `uint32_t` / `size_t` conversions,
or old additions that may potentially overflow).

This CL was uploaded by git cl split.

R=ddorwin@chromium.org, dtrainor@chromium.org, evliu@google.com, mfoltz@chromium.org, muyaoxu@google.com, nhiroki@chromium.org

Bug: 40284755
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5461340
Commit-Queue: Łukasz Anforowicz <lukasza@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1289336}

+++ components/cast_streaming/browser/frame/stream_consumer.cc
@@ -8,6 +8,7 @@
 
 #include "base/containers/span.h"
 #include "base/logging.h"
+#include "base/numerics/safe_conversions.h"
 #include "base/task/sequenced_task_runner.h"
 #include "components/cast_streaming/browser/common/decoder_buffer_factory.h"
 #include "components/cast_streaming/common/public/features.h"
@@ -110,7 +111,7 @@ void StreamConsumer::OnPipeWritable(MojoResult result) {
   }
 
   base::span<uint8_t> span = data_wrapper_.Get();
-  uint32_t bytes_written = span.size();
+  size_t bytes_written = span.size();
   result = data_pipe_->WriteData(span.data(), &bytes_written,
                                  MOJO_WRITE_DATA_FLAG_NONE);
   if (result != MOJO_RESULT_OK) {
@@ -118,7 +119,7 @@ void StreamConsumer::OnPipeWritable(MojoResult result) {
     return;
   }
 
-  data_wrapper_.Consume(bytes_written);
+  data_wrapper_.Consume(base::checked_cast<uint32_t>(bytes_written));
   if (!data_wrapper_.empty()) {
     pipe_watcher_.ArmOrNotify();
     return;
@@ -195,7 +196,7 @@ void StreamConsumer::MaybeSendNextFrame() {
 
   // Write the frame's data to Mojo.
   span = data_wrapper_.Get();
-  uint32_t bytes_written = span.size();
+  size_t bytes_written = span.size();
   auto result = data_pipe_->WriteData(span.data(), &bytes_written,
                                       MOJO_WRITE_DATA_FLAG_NONE);
   if (result == MOJO_RESULT_SHOULD_WAIT) {
@@ -205,7 +206,7 @@ void StreamConsumer::MaybeSendNextFrame() {
     CloseDataPipeOnError();
     return;
   }
-  data_wrapper_.Consume(bytes_written);
+  data_wrapper_.Consume(base::checked_cast<uint32_t>(bytes_written));
 
   // Return the frame.
   is_read_pending_ = false;
