+++ components/openscreen_platform/tls_client_connection.h
@@ -35,7 +35,7 @@ class TlsClientConnection final : public openscreen::TlsConnection {
   openscreen::IPEndpoint GetRemoteEndpoint() const final;
 
   // The maximum size of the vector in any single Client::OnRead() callback.
-  static constexpr uint32_t kMaxBytesPerRead = 64 << 10;  // 64 KB.
+  static constexpr size_t kMaxBytesPerRead = 64 << 10;  // 64 KB.
 
  private:
   // Invoked by |receive_stream_watcher_| when the |receive_stream_|'s status
